fix backup problem for dot files

+++ install.sh
@@ -1,23 +1,29 @@
-#!/bin/bash -ex
+#!/bin/bash -e
 cd $( dirname ${BASH_SOURCE[0]} )
 
-# install dot files
-pushd dotfiles
+echo install dot files
+pushd dotfiles > /dev/null
     for file in *
     do
         if [ -f ~/.$file ]; then
-            mv ~/.$file ~/.$file.bk
+            if ! cmp -s $PWD/$file ~/.$file ; then
+                if [ -f ~/.$file.bk ]; then
+                    echo "sorry, ~/.$file.bk exist, can't backup"
+                    exit 1
+                fi 
+                mv ~/.$file ~/.$file.bk
+            fi
         fi 
         ln -fs $PWD/$file ~/.$file
     done
-popd
+popd > /dev/null
 
-# setup vim
+echo setup vim
 curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 vim +PlugInstall +qall
 
-# setup shell
+echo setup shell
 rm -f ~/.bashrc.sel
 echo add \"source $PWD/shell/bashrc.ubuntu\" in .bashrc
